+++ src/main/java/solutions/dmitrikonnov/einstufungstest/businesslayer/ETAufgabenRestricter.java
@@ -8,6 +8,6 @@ import java.util.Map;
 
 public interface ETAufgabenRestricter {
 
-    public List<ETAufgabe> restrict(List<ETAufgabe> selectedAndReshuffeled, Map<ETAufgabenNiveau,Integer> niveauToMax);
+    public List<ETAufgabe> restrict(List<ETAufgabe> selectedAndReshuffeled, Map<ETAufgabenNiveau,Short> niveauToMax);
     public List<ETAufgabe> restrict(List<ETAufgabe> selectedAndReshuffeled);
 }
